requirements.txt
pytorch==2.0.0
transformers==4.31.0
gradio
sentencepiece
protobuf
cpm-kernels
icetk